ENH: generalize the SOFB SelectionWidget and move it to the widgets subpackage
