+++ pyqt-apps/siriushla/as_ap_sofb/ioc_control/respmat_enbllist.py
@@ -4,258 +4,15 @@ from functools import partial as _part
 
 import numpy as np
 
-from qtpy.QtWidgets import QGridLayout, QHBoxLayout, QLabel, \
-    QSizePolicy, QScrollArea, QWidget, QPushButton, QDialog, QTabWidget
-from qtpy.QtCore import Qt, QRect, QPoint, Signal
-from qtpy.QtGui import QBrush, QColor, QPainter
+from qtpy.QtWidgets import QHBoxLayout, \
+    QSizePolicy, QWidget, QDialog, QTabWidget
 
 from pydm.widgets.base import PyDMWidget
 
 from siriushla.widgets import SiriusLedAlert, QLed, \
-    SiriusConnectionSignal as _ConnSignal
+    SiriusConnectionSignal as _ConnSignal, \
+    SelectionMatrixWidget as SelectionWidget
 from siriushla.as_ap_sofb.ioc_control.base import BaseObject, BaseWidget
-
-
-class SelectionWidget(QWidget):
-    """Widget to perform component selection.
-
-    Parameters
-    ----------
-    parent : QWidget, optional
-        The parent widget for the SelectionMatrixWidget.
-    title: str, optional
-        Widget title.
-    has_bothplanes: bool, optional
-        Whether to show button to send applyBothPlanesClicked signal.
-        Default: False.
-
-    Signals
-    -------
-    applyChangesClicked:
-        emitted when "Apply Changes" button is clicked
-    applyBothPlanesClicked:
-        emitted when "Apply Both Planes" button is clicked
-    """
-
-    applyChangesClicked = Signal()
-    applyBothPlanesClicked = Signal()
-
-    def __init__(self, parent=None, title='', has_bothplanes=False):
-        """Init."""
-        super().__init__(parent)
-
-        self.title = title
-        self.has_bothplanes = has_bothplanes
-
-        self.begin = QPoint()
-        self.end = QPoint()
-
-        self._top_headers, self._side_headers = self.get_headers()
-        self._widgets = self.get_widgets()
-        self._positions = self.get_positions()
-        self._top_header_wids, self._side_header_wids = list(), list()
-        self._setupUi()
-
-    def _setupUi(self):
-        lay = QGridLayout(self)
-
-        lab = QLabel(self.title, self, alignment=Qt.AlignCenter)
-        lab.setStyleSheet("font-weight: bold;")
-        lay.addWidget(lab, 0, 0, 1, 1)
-
-        # scroll area + widgets matrix
-        scr_ar = QScrollArea(self)
-        scr_ar_wid = QWidget()
-        scr_ar_wid.setObjectName('scrollarea')
-        scr_ar_wid.setStyleSheet(
-            '#scrollarea {background-color: transparent;}')
-        scr_ar.setWidgetResizable(True)
-        scr_ar.setWidget(scr_ar_wid)
-        lay.addWidget(scr_ar, 1, 0, 1, 1)
-
-        glay = QGridLayout(scr_ar_wid)
-        glay.setContentsMargins(0, 0, 0, 0)
-        for i, head in enumerate(self._top_headers):
-            head_wid = QPushButton(head, self)
-            head_wid.setStyleSheet('min-width:2em;')
-            head_wid.clicked.connect(
-                _part(self.selectWidgetsAt, i, isrow=False))
-            self._top_header_wids.append(head_wid)
-            glay.addWidget(head_wid, 0, i+1)
-        for i, head in enumerate(self._side_headers):
-            head_wid = QPushButton(head, self)
-            head_wid.setStyleSheet('min-width:2em;')
-            head_wid.clicked.connect(
-                _part(self.selectWidgetsAt, i, isrow=True))
-            self._side_header_wids.append(head_wid)
-            glay.addWidget(head_wid, i+1, 0)
-        for i, wid in enumerate(self._widgets):
-            pos = self._positions[i]
-            glay.addWidget(wid, pos[0]+1, pos[1]+1)
-            led = wid.findChild(QLed)
-            if not led:
-                continue
-            led.clicked.connect(led.toggleSelected)
-
-        # action buttons
-        self.btn_unsel_all = QPushButton('Undo Selection')
-        self.btn_unsel_all.clicked.connect(self.undoItemsSelection)
-        self.btn_dsbl_all = QPushButton('Disable All')
-        self.btn_dsbl_all.clicked.connect(_part(self.toogleAllItems, False))
-        self.btn_enbl_all = QPushButton('Enable All')
-        self.btn_enbl_all.clicked.connect(_part(self.toogleAllItems, True))
-        self.btn_send = QPushButton('Apply Changes')
-        self.btn_send.clicked.connect(self.applyChangesClicked.emit)
-        if self.has_bothplanes:
-            self.btn_send_otpl = QPushButton('Apply Both Planes')
-            self.btn_send_otpl.clicked.connect(
-                self.applyBothPlanesClicked.emit)
-
-        hlay = QHBoxLayout()
-        hlay.addStretch()
-        hlay.addWidget(self.btn_unsel_all)
-        hlay.addStretch()
-        hlay.addWidget(self.btn_dsbl_all)
-        hlay.addStretch()
-        hlay.addWidget(self.btn_enbl_all)
-        hlay.addStretch()
-        hlay.addWidget(self.btn_send)
-        hlay.addStretch()
-        if self.has_bothplanes:
-            hlay.addWidget(self.btn_send_otpl)
-        lay.addLayout(hlay, 2, 0, 1, 1)
-
-        lay.setSizeConstraint(lay.SetMinimumSize)
-
-    def paintEvent(self, _):
-        """Paint event to draw selection rectangle."""
-        if self.begin == self.end:
-            return
-        qp = QPainter(self)
-        br = QBrush(QColor(100, 10, 10, 40))
-        qp.setBrush(br)
-        qp.drawRect(QRect(self.begin, self.end))
-
-    def mousePressEvent(self, event):
-        """Mouse press event."""
-        self.begin = event.pos()
-        self.end = event.pos()
-        self.update()
-
-    def mouseMoveEvent(self, event):
-        """Mouse move event."""
-        self.end = event.pos()
-        self.update()
-
-    def mouseReleaseEvent(self, event):
-        """Mouse release event."""
-        self.end = event.pos()
-        self.selectItems()
-        self.begin = event.pos()
-        self.update()
-
-    def selectItems(self):
-        """Select items."""
-        for i, wid in enumerate(self._widgets):
-            if not wid.isVisible():
-                continue
-            led = wid.findChild(QLed)
-            if not led:
-                continue
-            pos = led.mapTo(self, led.pos())
-            sz = led.size()
-            x1 = pos.x()+sz.width()/2 > self.begin.x()
-            x2 = pos.x()+sz.width()/2 > self.end.x()
-            y1 = pos.y()+sz.height()/2 > self.begin.y()
-            y2 = pos.y()+sz.height()/2 > self.end.y()
-            if x1 != x2 and y1 != y2:
-                led.toggleSelected()
-
-    def undoItemsSelection(self):
-        for wid in self._widgets:
-            led = wid.findChild(QLed)
-            if not led:
-                continue
-            led.setSelected(False)
-
-    def toogleAllItems(self, value):
-        for wid in self._widgets:
-            led = wid.findChild(QLed)
-            if not led:
-                continue
-            if not (bool(led.state) != value):
-                led.setSelected(True)
-
-    def selectWidgetsAt(self, idx, isrow=False):
-        for i, wid in enumerate(self._widgets):
-            row, col = self._positions[i]
-            led = wid.findChild(QLed)
-            if not led:
-                continue
-            if isrow and row == idx:
-                led.toggleSelected()
-            if not isrow and col == idx:
-                led.toggleSelected()
-
-    # --- properties ---
-
-    @property
-    def headers(self):
-        """Return top and side header text lists, respectively."""
-        return self._top_headers, self._side_headers
-
-    @property
-    def header_widgets(self):
-        """Return top and side header widget lists, respectively."""
-        return self._top_header_wids, self._side_header_wids
-
-    @property
-    def widgets(self):
-        """Return widget list."""
-        return self._widgets
-
-    @property
-    def positions(self):
-        """Return widget position list."""
-        return self._positions
-
-    # --- specific methods, should be implemented in derivation ---
-
-    def get_headers(self):
-        """
-        Should be implemented in class derivation.
-
-        Return
-        ------
-        top_headers: tuple or list
-            A list of strings for top headers of the selection matrix widget.
-        side_headers: tuple or list
-            A list of strings for side headers of the selection matrix widget.
-        """
-        raise NotImplementedError
-
-    def get_widgets(self):
-        """
-        Should be implemented in class derivation.
-
-        Return
-        ------
-        widgets: tuple or list
-            A tuple or list of widgets to be put in matrix.
-        """
-        raise NotImplementedError
-
-    def get_positions(self):
-        """
-        Should be implemented in class derivation.
-
-        Return
-        ------
-        positions: tuple or list
-            A tuple or list of layout positions for each widget
-            returned by get_widgets.
-        """
-        raise NotImplementedError
 
 
 class SingleSelMatrix(BaseObject, SelectionWidget, PyDMWidget):
