+++ pyqt-apps/siriushla/widgets/__init__.py
@@ -26,3 +26,4 @@ from .process_image import SiriusProcessImage
 from .detachable_tabwidget import DetachableTabWidget
 from .pushbutton import SiriusPushButton
 from .waveformtable import SiriusWaveformTable
+from .selection_matrix import SelectionMatrixWidget
